Include TGGraphLayout in the source tree.
I believe this is permitted (the TGGraphLayout JAR was released under the Apache license). However, it does mean that the jgrapht-touchgraph demo will not be accepted by central or by other repositories that ban <repository> references - do not run releases with -Ptouchgraph!

The '3rdparty' repository tree was populated using

	mvn deploy:deploy-file \
		-DgroupId=com.touchgraph \
		-DartifactId=TGGraphLayout \
		-Dfile=TGGraphLayout.jar \
		-Dpackaging=jar \
		-Dversion=UNKNOWN \
		-Durl=file:/path/to/jgrapht/jgrapht-touchgraph/3rdparty

I've also included the license file.